services/finance-jobs-worker/requirements.txt
fastapi>=0.110
uvicorn>=0.29
supabase>=2.4
pydantic>=2.6
pydantic-settings>=2.2
google-cloud-tasks>=2.16
